requirements.txt
scipy>=0.18.1
numpy>=1.11.2
pytest>=3.0
matplotlib
progressbar2
pandas
biopython
re
sklearn
